Create Four-Course.tf
In this example, we have deployed four resources to four different platforms: AWS, Kubernetes, Atlas Mongo DB, and Azure. We first created our EKS Kubernetes cluster. Then we deployed a workload to that Kubernetes cluster. We created a Mongo DB cluster in Atlas that our workload could connect to. Finally, we created a DNS entry in Azure DNS to point to our workload in EKS